backend/infoauto/requirements/base.txt
# Basic
# ------------------------------------------------------------------------------
pytz==2018.5  # https://github.com/stub42/pytz
python-slugify==1.2.6  # https://github.com/un33k/python-slugify
Pillow==5.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.3.0  # https://github.com/hynek/argon2_cffi

# Django
# ------------------------------------------------------------------------------
django==2.2.0  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.37.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
rest_condition==1.0.3
django-filter==2.0.0
drf-writable-nested==0.5.0
djangorestframework-xml==1.3.0

# Swagger
# ------------------------------------------------------------------------------
drf-yasg==1.13.0
flex==6.13.2
swagger-spec-validator==2.3.1
jsonschema==2.6.0

# MySQL
# ------------------------------------------------------------------------------
mysqlclient==1.3.13

# OpenPyXL - A Python library to read/write Excel 2010 xlsx/xlsm files
# ------------------------------------------------------------------------------
openpyxl==2.5.8

# Django historic for models
# ------------------------------------------------------------------------------
django-simple-history==2.4.0

# Memcached
# ------------------------------------------------------------------------------
python-memcached==1.59

sqlparse==0.3.0

# WebDAV client - Provides easy and convenient work with WebDAV-servers (Dropbox, Google Drive, Box, 4shared, etc.)
# ------------------------------------------------------------------------------
# $ sudo apt-get install libxml2-dev libxslt-dev python-dev
# $ sudo apt-get install libcurl4-openssl-dev python-pycurl
# $ sudo easy_install webdavclient

webdavclient3==0.10

# Celery
# ------------------------------------------------------------------------------
tornado==5.1.1
celery==4.3.1  # pyup: <5.0  # https://github.com/celery/celery
billiard==3.6.0.0
kombu==4.5.0
flower==0.9.3  # https://github.com/mher/flower
redis==3.3.8   # https://github.com/andymccurdy/redis-py
django-celery-beat==1.5.0
django-celery-results==1.1.2

# Channels (WebSocket)
# ------------------------------------------------------------------------------
channels==2.1.7
daphne==2.2
channels_redis==2.3.3

django-picklefield==2.0.0
requests==2.22.0


# Django
# ------------------------------------------------------------------------------
docutils==0.11
django-storages[boto3]==1.7.1  # https://github.com/jschneier/django-storages
django-anymail[mailgun]==4.2  # https://github.com/anymail/django-anymail
beautifulsoup4==4.9.0


# memcached
python-memcached==1.59
tqdm==4.39.0

XlsxWriter==1.2.8
xlrd==1.2.0

# HubSpot
# ------------------------------------------------------------------------------
hubspot3==3.2.50